Extraction of Amazon Best Selling Books in 2023
The is a file showing data scraped from Amazon website for best selling books in 2023  on from the website. The data was unstructured and dirty when exported to this Excel after being scraped as seen on the first page of the file labeled scraped data. It had to be cleaned using the necessary Excel tools as shown on the second page of the file labeled cleaned data